pyproject.toml
[tool.poetry]
name = "python_project_template"
version = "0.1.0"
description = ""
authors = ["davidtnfsh <davidtnfsh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.26.0"
apache-airflow = "1.10.13"
searchconsole = "^0.0.4"
pydantic = "^1.8.2"
google-cloud-bigquery = {version = "^2.24.0", python = ">=3.7.1,<3.10"}
pandas = "^1.3.2"
pyarrow = "^5.0.0"
StrEnum = "^0.4.6"
python-fb-page-insights-client = {git = "https://github.com/pycontw/python-fb-page-insights-client.git", rev = "bd221ff"}
Flask-OAuthlib = "^0.9.6"
Flask-OpenID = "1.3.0"
pygsheets = "^2.0.6"

[tool.poetry.dev-dependencies]
safety = "^1.9.0"
bandit = "^1.6.2"
pytest-cov = "^2.8.1"
black = "^19.10b0"
mypy = "^0.770"
pytest = "^5.4.1"
flake8 = "^3.8.3"
isort = "^5.6.4"

[build-system]
requires = ["setuptools", "poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"


[tool.coverage]
    [tool.coverage.report]
    show_missing = true
    exclude_lines = [
        # Have to re-enable the standard pragma
        'pragma: no cover',

        # Don't complain about missing debug-only code:
        'def __repr__',
        'if self\.debug',

        # Don't complain if tests don't hit defensive assertion code:
        'raise AssertionError',
        'raise NotImplementedError',

        # Don't complain if non-runnable code isn't run:
        'if 0:',
        'if __name__ == .__main__.:'
    ]
    omit = [
        'env/*',
        'venv/*',
        '*/virtualenv/*',
        '*/virtualenvs/*',
        '*/tests/*'
    ]
